improve the organization of process manager
