+++ src/tracer/src/extracts/process/process_manager/mod.rs
@@ -1,7 +1,9 @@
-pub mod extract_process_data;
+pub mod handlers;
+pub mod logger;
+pub mod matcher;
 pub mod process_manager;
-pub mod process_utils;
-pub mod types;
+pub mod state;
+pub mod system_refresher;
 
 // Re-export the main ProcessManager for convenience
 pub use process_manager::ProcessManager;
